refactor list iocs action response
